Modified the Drawer navigation to better organize the file structure.
