+++ frontend/_tests_/CampusMap.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import CampusMap from "@/app/OutdoorNav/CampusMap";
+import CampusMap from "@/app/screens/OutdoorNav/CampusMap";
 
 test("Campus Map renders correctly", () => {
   const tree = renderer.create(<CampusMap />).toJSON();
