+++ frontend/app/components/ui/AppLogo.tsx
@@ -2,15 +2,15 @@
 //Renders the Campus Pilot logo
 import { View, Text, Image, StyleSheet } from "react-native";
 import React from "react";
-import useTheme from "../hooks/useTheme";
+import useTheme from "../../hooks/useTheme";
 
-export default function appLogo() {
+export default function AppLogo() {
   const { theme, toggleTheme } = useTheme(); // Use the custom hook
   return (
     <View>
       <Image
         style={styles.logo}
-        source={require("../../assets/images/logo.png")}
+        source={require("../../../assets/images/logo.png")}
         tintColor={theme === "dark" ? "#000" : "#fff"} // Invert the color based on theme
       />
     </View>
