+++ frontend/tsconfig.json
@@ -14,5 +14,5 @@
     "**/*.tsx",
     ".expo/types/**/*.ts",
     "expo-env.d.ts"
-, "app/OutdoorNav/CampusMaptsx", "app/components/ThemeToggle.tsx", "app/styles.tsx"  ]
+, "app/screens/OutdoorNav/CampusMaptsx", "app/components/ui/ThemeToggle.tsx", "app/styles.tsx"  ]
 }
